requirements.txt
ccxt
numpy
pandas
scikit-learn
tensorflow
python-telegram-bot
python-dotenv
